Key Setter Component
The Key Setter field has been updated to select field, for which we are setting up all possible paths that a data model has.
Also, we can use choicejs to our advantage (which is provied by FORM.IO out of the box) for getting the IntelliSense type of functionality.
For which we extract all the possible paths from a data model and pass those values to the KeySetter, where searching for any Path is made easier